+++ examples/dogsvscats/demo/demo.go
@@ -37,8 +37,10 @@ import (
 	"github.com/gomlx/gomlx/ml/train/losses"
 	"github.com/gomlx/gomlx/ml/train/metrics"
 	"github.com/gomlx/gomlx/ml/train/optimizers"
+	"github.com/gomlx/gomlx/models/inceptionv3"
 	"github.com/gomlx/gomlx/types/slices"
 	"github.com/gomlx/gomlx/types/tensor"
+	timage "github.com/gomlx/gomlx/types/tensor/image"
 	"log"
 	"os"
 	"path"
@@ -56,13 +58,16 @@ var (
 	flagCheckpointKeep = flag.Int("checkpoint_keep", 10, "Number of checkpoints to keep, if --checkpoint is set.")
 
 	// Training hyperparameters:
-	flagOptimizer        = flag.String("optimizer", "adamw", fmt.Sprintf("Optimizer, options: %q", slices.SortedKeys(optimizers.KnownOptimizers)))
-	flagNumSteps         = flag.Int("steps", 2000, "Number of gradient descent steps to perform")
-	flagBatchSize        = flag.Int("batch", dogsvscats.DefaultConfig.BatchSize, "Batch size for training")
-	flagLearningRate     = flag.Float64("learning_rate", 0.0001, "Initial learning rate.")
-	flagL2Regularization = flag.Float64("l2_reg", 0, "L2 regularization on kernels. It doesn't interact well with --batch_norm.")
-	flagNormalization    = flag.String("norm", "layer", fmt.Sprintf("Type of layer normalization to use. Valid values: %q.", slices.SortedKeys(layers.KnownNormalizers)))
-	flagEval             = flag.Bool("eval", true, "Whether to evaluate trained model on test data in the end.")
+	flagModelType           = flag.String("model", "cnn", "Model types: \"cnn\" or \"inception\"")
+	flagOptimizer           = flag.String("optimizer", "adamw", fmt.Sprintf("Optimizer, options: %q", slices.SortedKeys(optimizers.KnownOptimizers)))
+	flagNumSteps            = flag.Int("steps", 2000, "Number of gradient descent steps to perform")
+	flagBatchSize           = flag.Int("batch", dogsvscats.DefaultConfig.BatchSize, "Batch size for training")
+	flagLearningRate        = flag.Float64("learning_rate", 0.0001, "Initial learning rate.")
+	flagL2Regularization    = flag.Float64("l2_reg", 0, "L2 regularization on kernels. It doesn't interact well with --batch_norm.")
+	flagNormalization       = flag.String("norm", "layer", fmt.Sprintf("Type of layer normalization to use. Valid values: %q.", slices.SortedKeys(layers.KnownNormalizers)))
+	flagEval                = flag.Bool("eval", true, "Whether to evaluate trained model on test data in the end.")
+	flagInceptionPreTrained = flag.Bool("pretrained", true, "If using inception model, whether to use the pre-trained weights to transfer learn")
+	flagInceptionFineTuning = flag.Bool("finetuning", true, "If using inception model, whether to fine-tune the inception model")
 
 	// Flat part of model, after convolutions and models being flattened:
 	flagNumHiddenLayers = flag.Int("hidden_layers", 3, "Number of hidden layers, stacked with residual connection.")
@@ -101,6 +106,9 @@ func main() {
 	config.AngleStdDev = *flagAngleStdDev
 	config.FlipRandomly = *flagFlipRandomly
 	config.BatchSize = *flagBatchSize
+	if *flagModelType == "inception" {
+		config.ModelImageSize = inceptionv3.MinimumImageSize
+	}
 	if *flagPreGenerate {
 		dogsvscats.PreGenerate(config, *flagPreGenEpochs)
 		return
@@ -137,9 +145,21 @@ func trainModel(config *dogsvscats.Configuration) {
 		AssertNoError(err)
 	}
 
+	// Select the model type we are using:
+	var modelFn func(ctx *context.Context, spec any, inputs []*Node) []*Node
+	switch *flagModelType {
+	case "cnn":
+		modelFn = CnnModelGraph
+	case "inception":
+		modelFn = InceptionV3ModelGraph
+		if *flagInceptionPreTrained {
+			AssertNoError(inceptionv3.DownloadAndUnpackWeights(*flagDataDir))
+		}
+	}
+
 	// Create a train.Trainer: this object will orchestrate running the model, feeding
 	// results to the optimizer, evaluating the metrics, etc. (all happens in trainer.TrainStep)
-	trainer := train.NewTrainer(manager, ctx, modelGraph,
+	trainer := train.NewTrainer(manager, ctx, modelFn,
 		losses.BinaryCrossentropyLogits,
 		optimizers.MustOptimizerByName(*flagOptimizer),
 		[]metrics.Interface{movingAccuracyMetric}, // trainMetrics
@@ -162,7 +182,7 @@ func trainModel(config *dogsvscats.Configuration) {
 	_, err := loop.RunSteps(trainDS, *flagNumSteps)
 	AssertNoError(err)
 
-	// Finally print an evaluation on train and test datasets.
+	// Finally, print an evaluation on train and test datasets.
 	if *flagEval {
 		fmt.Println()
 		err = commandline.ReportEval(trainer, trainEvalDS, validationEvalDS)
@@ -198,8 +218,9 @@ func normalizeFeatures(ctx *context.Context, x *Node) *Node {
 	return g.InvalidNode()
 }
 
-// modelGraph builds the model for our demo. It returns the logits, not the predictions, which works with most losses.
-func modelGraph(ctx *context.Context, spec any, inputs []*Node) []*Node {
+// CnnModelGraph builds the CNN model for our demo. It returns the logits, not the predictions, which works with
+// most losses.
+func CnnModelGraph(ctx *context.Context, spec any, inputs []*Node) []*Node {
 	_ = spec // Not needed.
 	x := inputs[0]
 	g := x.Graph()
@@ -215,6 +236,7 @@ func modelGraph(ctx *context.Context, spec any, inputs []*Node) []*Node {
 		logits = layers.Convolution(ctx, logits).Filters(filterSize).KernelSize(3).PadSame().Done()
 		logits = layers.Relu(logits)
 		logits = normalizeImage(ctx, logits)
+
 		if convIdx > 0 {
 			logits = Add(logits, residual)
 		}
@@ -222,8 +244,14 @@ func modelGraph(ctx *context.Context, spec any, inputs []*Node) []*Node {
 		AssertNoError(g.Error())
 	}
 
-	// Flatten resulting image, and treat the convolved logits as tabular.
+	// Flatten the resulting image, and treat the convolved logits as flattened.
 	logits = Reshape(logits, batchSize, -1)
+	logits = FnnOnTop(ctx, logits)
+	return []*Node{logits}
+}
+
+// FnnOnTop adds a feedforward neural network on top of the CNN layer.
+func FnnOnTop(ctx *context.Context, logits *Node) *Node {
 	for ii := 0; ii < *flagNumHiddenLayers; ii++ {
 		ctx := ctx.In(fmt.Sprintf("dense_%d", ii))
 		residual := logits
@@ -237,5 +265,33 @@ func modelGraph(ctx *context.Context, spec any, inputs []*Node) []*Node {
 	}
 	logits = layers.Relu(logits)
 	logits = layers.DenseWithBias(ctx.In("readout"), logits, 1)
+	return logits
+}
+
+// InceptionV3ModelGraph uses an optionally pre-trained inception model.
+//
+// Results on validation after training 1000 steps, inception pre-trained weights used:
+// - no scaling (from 0.0 to 1.0): 90.4% accuracy
+// - with Keras scale (from -1.0 to 1.0): 90.6% accuracy
+//
+// Results if we don't use the pre-trained weights (it can probably get much better with more training):
+// - no scaling (from 0.0 to 1.0): 62.5% accuracy
+// - with Keras scale (from -1.0 to 1.0): 61.8% accuracy
+func InceptionV3ModelGraph(ctx *context.Context, spec any, inputs []*Node) []*Node {
+	_ = spec           // Not needed.
+	image := inputs[0] // Images scaled from 0.0 to 1.0
+	channelsConfig := timage.ChannelsLast
+	image = inceptionv3.PreprocessImage(image, 1.0, channelsConfig)
+
+	var preTrainedPath string
+	if *flagInceptionPreTrained {
+		preTrainedPath = *flagDataDir
+	}
+	logits := inceptionv3.BuildGraph(ctx, image).PreTrained(preTrainedPath).
+		SetPooling(inceptionv3.MaxPooling).Trainable(*flagInceptionFineTuning).Done()
+	if !*flagInceptionFineTuning {
+		logits = StopGradient(logits) // We don't want to train the inception model.
+	}
+	logits = FnnOnTop(ctx, logits)
 	return []*Node{logits}
 }